# Conflicts:
#	SP4/Assets/Resources/Heroes/CrimsonKnight.cs

+++ SP4/Assets/Scripts/Gameplay/Hero.cs
@@ -132,7 +132,7 @@ public class Hero : MonoBehaviour {
 		}
 	}
 
-	public virtual void SetAttack(int newAtk)
+	public virtual void SetAttack(float newAtk)
 	{
 		Attack = newAtk;
 	}
